update customized for release v2.7.0

+++ installer/wecube-system-settings/configure-artifacts.sh
@@ -6,31 +6,38 @@ SYS_SETTINGS_ENV_FILE=$1
 
 source $SYS_SETTINGS_ENV_FILE
 
-([ -z "$ARTIFACTS_COS_REGION" ] || [ -z "$ARTIFACTS_COS_BUCKET" ] || [ -z "$ARTIFACTS_COS_OBJECTS" ]) && \
-  echo "No artifacts are specified and skipped uploading." && exit 0
-
-echo "Downloading artifacts from TencentCloud COS..."
-PYTHON_SCRIPT_FILE=$(realpath "./download-objects-in-bucket.py")
 DOWNLOAD_DIR=$(realpath "./download")
 mkdir -p "$DOWNLOAD_DIR"
 
-PIP_MIRROR_PARAM=""
-if [ "$USE_MIRROR_IN_MAINLAND_CHINA" == "true" ]; then
-	echo 'Using mirror for pip index in Mainland China.'
-	PIP_MIRROR_PARAM="-i https://pypi.tuna.tsinghua.edu.cn/simple"
-fi
-read -d '' SHELL_CMD <<-EOF || true
-	pip install $PIP_MIRROR_PARAM -U crcmod cos-python-sdk-v5 && \
-	python $PYTHON_SCRIPT_FILE \
-		$ARTIFACTS_COS_SECRETID $ARTIFACTS_COS_SECRETKEY \
-		$ARTIFACTS_COS_REGION $ARTIFACTS_COS_BUCKET \"${ARTIFACTS_COS_OBJECTS[*]}\" \
-		$DOWNLOAD_DIR
-EOF
-docker run --rm -t \
-	-v "$PYTHON_SCRIPT_FILE:$PYTHON_SCRIPT_FILE" \
-	-v "$DOWNLOAD_DIR:$DOWNLOAD_DIR" \
-	python:3 /bin/sh -c "$SHELL_CMD"
+for ARTIFACT_URL in "${ARTIFACTS_PKGS[@]}"; do
+	ARTIFACT_FILE="$DOWNLOAD_DIR/${ARTIFACT_URL##*'/'}"
+	echo "Fetching artifact from $ARTIFACT_URL"
+	../curl-with-retry.sh -fL $ARTIFACT_URL -o $ARTIFACT_FILE
+done
 
+if [ -n "$ARTIFACTS_COS_REGION" ] && [ -n "$ARTIFACTS_COS_BUCKET" ] && [ -n "$ARTIFACTS_COS_OBJECTS" ]; then
+	echo -e "\nFetching artifacts from TencentCloud COS..."
+	PYTHON_SCRIPT_FILE=$(realpath "./download-objects-in-bucket.py")
+
+	PIP_MIRROR_PARAM=""
+	if [ "$USE_MIRROR_IN_MAINLAND_CHINA" == "true" ]; then
+		echo 'Using mirror for pip index in Mainland China.'
+		PIP_MIRROR_PARAM="-i https://pypi.tuna.tsinghua.edu.cn/simple"
+	fi
+	read -d '' SHELL_CMD <<-EOF || true
+		pip install $PIP_MIRROR_PARAM -U crcmod cos-python-sdk-v5 && \
+		python $PYTHON_SCRIPT_FILE \
+			$ARTIFACTS_COS_SECRETID $ARTIFACTS_COS_SECRETKEY \
+			$ARTIFACTS_COS_REGION $ARTIFACTS_COS_BUCKET \"${ARTIFACTS_COS_OBJECTS[*]}\" \
+			$DOWNLOAD_DIR
+	EOF
+	docker run --rm -t \
+		-v "$PYTHON_SCRIPT_FILE:$PYTHON_SCRIPT_FILE" \
+		-v "$DOWNLOAD_DIR:$DOWNLOAD_DIR" \
+		python:3 /bin/sh -c "$SHELL_CMD"
+fi
+
+echo -e "\nVerifying MD5 checksum..."
 find "$DOWNLOAD_DIR" -type f | while read FILE; do
 	BASE_NAME=$(basename $FILE)
 	DIR_NAME=$(dirname $FILE)
@@ -44,7 +51,7 @@ find "$DOWNLOAD_DIR" -type f | while read FILE; do
 	fi
 done
 
-echo "Uploading artifacts..."
+echo -e "\nUploading artifacts to plugin storage..."
 SHELL_SCRIPT_FILE=$(realpath "./upload-all-files-in-dir-to-minio.sh")
 docker run --rm -t \
 	-v "$DOWNLOAD_DIR:$DOWNLOAD_DIR" \
